+++ packages/engine/Source/Scene/Model/createVectorTileBuffersFromModelComponents.js
@@ -575,7 +575,7 @@ function appendNodeToBuffers(content, node, parentTransform, result) {
     // tileset cannot know the number of other tilesets it overlaps, or the number of collections within those
     // tilesets, so (if z-fighting is a concern) users must assign tileset z-index values in increments >=N, where N is
     // the number of collections/nodes/layers expected in each tileset.
-    const tilesetZIndex = content.tileset._zIndex ?? 0;
+    const tilesetZIndex = content.tileset._vectorZIndex ?? 0;
     const collectionIndex = result.collections.length;
 
     const stats = gatherPrimitiveStats(primitive);